Safari-611.1.10/LayoutTests/platform/win/platform/mac/accessibility/form-control-value-settable-expected.txt



This tests whether AXValue is writable for various form controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


text1
PASS document.activeElement == element1 is true
text2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is true
PASS axElement2.isAttributeSettable('AXValue') is true

number1
PASS document.activeElement == element1 is true
number2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is true
PASS axElement2.isAttributeSettable('AXValue') is true

textarea1
PASS document.activeElement == element1 is true
textarea2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is true
PASS axElement2.isAttributeSettable('AXValue') is true

slider1
PASS document.activeElement == element1 is true
slider2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is true
PASS axElement2.isAttributeSettable('AXValue') is true

checkbox1
PASS document.activeElement == element1 is true
checkbox2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is false
PASS axElement2.isAttributeSettable('AXValue') is true

radio1
PASS document.activeElement == element1 is true
radio2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is false
PASS axElement2.isAttributeSettable('AXValue') is true

submit1
PASS document.activeElement == element1 is true
submit2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is false
PASS axElement2.isAttributeSettable('AXValue') is true

combobox1
PASS document.activeElement == element1 is true
combobox2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is false
PASS axElement2.isAttributeSettable('AXValue') is true

listbox1
PASS document.activeElement == element1 is true
listbox2
PASS document.activeElement == element2 is true
PASS axElement1.isAttributeSettable('AXValue') is false
PASS axElement2.isAttributeSettable('AXValue') is true

PASS successfullyParsed is true

TEST COMPLETE
